requirements.txt
django==3.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
opencv-python==4.2.0.34
